Added precharge.c and precharge.h, added structures to timer.h for timing transmit events, moved the functionality of send_can() into handle_can(), changed handle_ass() to handle_pio(), moved bms state and ass state to precharge.h
precharge.c has the precharge function now along with the bms state and ass state. pio.c is still used to set the state of the ass relay. Two structures were added to timer.c called tx_ready and tx_time which are used to time transmitting events to the data logger, the battery, and the inverters. TODO: I have not written the code to transmit commands to the inverters, and all of the code I wrote this morning needs to be checked for bugs.