UART interrupts now scan for a \n character at the end of the serial queue  so that the uart_char_rx() function is only called once per message.
Added UART1_Peek and UART2_Peek to plib_uart1.c and plib_uart2.c, which check the last character in the UART buffer without reading. Used this function to make the interrupts function better - uart interrupts only call the uart_rx_char() function once per message by using UART_Peek to check whether a \n has been sent.  This was found to be necessary because if uart_rx_char is called on every 1 byte interrupt, it is possible to skip characters in the message. Also changed the interrupt handler so that it clears UART errors.