Added reading pack current to can.c and precharge.h, added delayed ass loop opening to pio.c
To ensure we don't damage the contactors unnecessarily, the VCU will start a 300ms timer when it detects a condition requiring that we open the ASS loop. If the VCU sees a message from the BMS indicating that the pack current is <1A, it will open the ASS loop immediately, otherwise the ASS loop opens when the 300ms timer expires.