Added support for TS start button, made changes to precharge logic based around TS button, changed car_control.ready_to_drive flag so that ready to drive implies that precharge is ready
The VCU reads a can message ID 0x500 from the dash to get the ts button state. (Logic is done separately by the dash to latch the ts button state so we can read it directly on the vcu). In pio.c, the ts button state is used directly to open or close the ASS relay. Thus, the TS button has the same effect as the TSMS used to have, so the precharge code can be changed minimally. Additional changes are made in precharge.c so that using the ignition switch can result in a safe reset of the precharge logic i.e. the precharge state returns to PC_TS_OFF. Additionally, logic was added to can.c so that car_control.ready_to_drive is only true when precharge is finished, and the driver can only set ready to drive when precharge is finished.