Cleanup of unnecessary code, added a pause function to timer.c
ass_flag_pedal_invalid is unused, we are instead relying on the timeout flag for invalid pedal measurements. So, the flag has been removed from pio.c and from the status message in can.c. Also, a pause function was added which takes a number of ms as an argument and pauses execution for that time, using while(!has_delay_passed());.