Untested code added to precharge.c that writes desired parameters to the inverters on finishing precharge.
This code is to be tested in a separate branch so that testing can safely be completed on a power supply instead of the battery. Also found a mistake in pio.c - in the area where we latch the EMSDC, we set bms.ams_precharge_enabled rather than bms.precharge_enable, which is the correct flag for the vcu to set. (bms.precharge_enable is what we send, bms.ams_precharge_enabled is what we receive.)