Made changes so that configuration can occur on startup and IDs are extracted from inverters
Increased the size of the buffers substantially and added code to the inverter parser to look for an ID in messages starting with "*". Added an additional mode of reading to uart0.c, where characters are read in 1199 character chunks until comms are registered as active. This means that the startup message all gets sent to the parser as a single message, comms are registered as active, and then we switch back to normal reading. Added an additional state to precharge, which is the WRITE_INVERTER_PARAMETERS state. This state goes between WAIT_FOR_FINAL_VOLTAGE and READY. During this state, we hand over control to a mini-state machine which writes each inverter parameter in turn, waiting 100ms in between each parameter write.