Changes made for startup sequence to work properly
Changes include: adding data structure instead of definitions to store the parameters that we write to the inverters on startup. Adding a CAN message that tells the dash what parameters we write. Getting rid of id checks at the beginning (it may be possible to bring them back in a relaxed form). Added a flag that tells the car we are currently in parameter writing mode, so we can't enter ready-to-drive.